# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Unified Service Scheduler API
  description: >
    Dealership appointment scheduler (Keyloop assessment, Scenario A).
    Auto-assigns a qualified technician and free service bay for a
    requested vehicle/service/dealership/time window, guaranteeing no two
    confirmed appointments overlap on the same technician or bay (enforced
    via MySQL pessimistic locking — see docs/design.md §3).
  version: "1.0.0"
servers:
  - url: http://localhost:8080
    description: Local (docker compose up --build)

paths:
  /appointments:
    post:
      summary: Book an appointment
      description: >
        Auto-assigns the first qualified technician (specialty matches
        serviceType, same dealershipId) and the first free service bay at
        that dealership for the requested window. Caller does not choose
        technicianId/serviceBayId directly.
      operationId: createAppointment
      parameters:
        - name: Idempotency-Key
          in: header
          required: false
          description: >
            Optional client-supplied key. A second concurrent request with
            the same key is rejected with 409 before it reaches booking
            logic. Omitting this header does not change behavior.
          schema:
            type: string
        - name: X-Correlation-Id
          in: header
          required: false
          description: Optional; generated by the server if omitted, echoed back on the response.
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAppointmentRequest"
            example:
              vehicleId: 1
              serviceType: OIL_CHANGE
              dealershipId: 1
              desiredStart: "2026-09-01T09:00:00"
              desiredEnd: "2026-09-01T10:00:00"
      responses:
        "201":
          description: Appointment confirmed
          headers:
            X-Correlation-Id:
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentResponse"
        "400":
          description: Validation failure (missing field, or desiredEnd not after desiredStart)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: vehicleId does not exist
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: >
            No qualified technician or free service bay available for the
            requested window, OR a request with the same Idempotency-Key
            is already in flight.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /appointments/{id}:
    get:
      summary: Get an appointment by id
      operationId: getAppointment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: Appointment found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AppointmentResponse"
        "404":
          description: No appointment with that id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /technicians/{id}/availability:
    get:
      summary: Get a technician's confirmed busy windows for a given day
      description: Manual/demo inspection endpoint, not a hard scenario requirement.
      operationId: getTechnicianAvailability
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
            format: int64
        - name: date
          in: query
          required: true
          schema:
            type: string
            format: date
      responses:
        "200":
          description: Busy windows for the day
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AvailabilityResponse"
        "404":
          description: No technician with that id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    CreateAppointmentRequest:
      type: object
      required: [vehicleId, serviceType, dealershipId, desiredStart, desiredEnd]
      properties:
        vehicleId:
          type: integer
          format: int64
        serviceType:
          type: string
          example: OIL_CHANGE
        dealershipId:
          type: integer
          format: int64
        desiredStart:
          type: string
          format: date-time
        desiredEnd:
          type: string
          format: date-time
          description: Must be after desiredStart.

    AppointmentResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
        vehicleId:
          type: integer
          format: int64
        vehicleVin:
          type: string
        customerId:
          type: integer
          format: int64
        technicianId:
          type: integer
          format: int64
        technicianName:
          type: string
        serviceBayId:
          type: integer
          format: int64
        bayNumber:
          type: string
        serviceType:
          type: string
        startTime:
          type: string
          format: date-time
        endTime:
          type: string
          format: date-time
        status:
          type: string
          enum: [CONFIRMED, CANCELLED]

    AvailabilityResponse:
      type: object
      properties:
        technicianId:
          type: integer
          format: int64
        date:
          type: string
          format: date
        busyWindows:
          type: array
          items:
            type: object
            properties:
              startTime:
                type: string
                format: date-time
              endTime:
                type: string
                format: date-time

    ErrorResponse:
      type: object
      properties:
        status:
          type: integer
        error:
          type: string
        message:
          type: string
        correlationId:
          type: string
        timestamp:
          type: string
          format: date-time
        fieldErrors:
          type: array
          items:
            type: string